requirements.txt
aiohttp==3.7.4.post0
anyio==4.9.0
async-timeout==3.0.1
attrs==25.3.0
audioop-lts==0.2.1
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
colorama==0.4.6
configparser==7.2.0
cryptography==45.0.4
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@8686b5d1bbc1d3c862292eb436ab630d6e9c9b53
discord-ui==5.1.6
discord.py==2.5.2
discord_utility_Fantasiuss==0.0.3
district42==1.9.0
docutils==0.21.2
dotenv==0.9.9
firebase-admin==6.9.0
google-api-core==2.24.2
google-api-python-client==2.173.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
legacy-cgi==2.6.3
markdown-it-py==3.0.0
mdurl==0.1.2
MIDIUtil==1.2.1
more-itertools==10.6.0
msgpack==1.1.1
multidict==6.2.0
nh3==0.2.21
niltype==1.0.2
oauthlib==3.2.2
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
revolt==1.9.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
sniffio==1.3.1
spacetimedb-sdk==0.7.0
th==0.4.1
twine==6.1.0
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
valera==1.9.0
websocket-client==1.8.0
websockets==10.4
yarl==1.18.3
